requirements.txt
asgiref==3.5.2
azure-core==1.24.1
azure-storage-blob==12.9.0
backports.zoneinfo==0.2.1
Brotli==1.0.9
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
cssselect2==0.6.0
Django==4.0.4
django-cors-headers==3.13.0
django-crontab==0.7.1
django-environ==0.8.1
django-storages==1.12
djangorestframework==3.13.1
fernet==1.0.1
firebase-admin==5.2.0
fonttools==4.33.3
geographiclib==1.52
geopy==2.2.0
google-api-core==2.8.2
google-api-python-client==2.53.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.4
idna==3.3
isodate==0.6.1
kafka-python==2.0.2
msgpack==1.0.4
msrest==0.7.1
numpy==1.23.1
oauthlib==3.2.0
pandas==1.4.3
Pillow==9.1.1
proto-plus==1.20.6
protobuf==3.20.1
psycopg2-binary==2.9.3
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydyf==0.2.0
PyJWT==2.4.0
pyparsing==3.0.9
pyphen==0.12.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-http-client==3.3.7

pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.2.0
stripe==5.2.0
tinycss2==1.1.1
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
weasyprint==55.0
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.1
